Dedicated package for synchronous generators
