+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/OmegaRef.java
@@ -10,7 +10,7 @@ import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.DynaWaltzParametersDatabase;
 import com.powsybl.dynawaltz.dynamicmodels.nonstaticref.network.BusModel;
-import com.powsybl.dynawaltz.dynamicmodels.staticref.generators.GeneratorSynchronousModel;
+import com.powsybl.dynawaltz.dynamicmodels.staticref.generators.synchronous.GeneratorSynchronousModel;
 import com.powsybl.dynawaltz.xml.ParametersXml;
 import com.powsybl.iidm.network.Generator;
 import org.apache.commons.lang3.tuple.Pair;
